SystemUI: ImageWallpaper: adapt the attrib_list to get egl_config with expected native format RGBx_8888.
Especially for 3288 device with Bifrost DDK of r12 and later.
Bifrost DDK r12 supports an new egl_config with native format of
RGBA_1010102,
ImageWallpaper before this commit would use that egl_config.
While, rk_vop does not support to process a graphics buffer with that
format,
hwc_composition could not be used, resulting in poor performance.

For the attrib_list modified, please refer to egl_1.4_spec, "3.4.1.2
Sorting of EGLConfigs", "3. Special: ".

Signed-off-by: chenzhen <chenzhen@rock-chips.com>
Signed-off-by: Bian Jin chen <kenjc.bian@rock-chips.com>
Change-Id: I035e479b03c8ecddec60a4b369bf0bde0f8d2b5f