bootanimation: set the attrib_list format to RGBx_8888.
Especially for 3288 device with Bifrost DDK of r12 and later.
Bifrost DDK r12 supports an new egl_config with native format of
RGBA_1010102, bootanim before this commit would use that egl_config.
While, rk_vop does not support to process a graphics buffer with that
format, hwc_composition could not be used, resulting in poor performance.

For the attrib_list modified, please refer to egl_1.4_spec, "3.4.1.2
Sorting of EGLConfigs", "3. Special: ".

Signed-off-by: chenzhen <chenzhen@rock-chips.com>
Signed-off-by: Zhixiong Lin <zhixiong.lin@rock-chips.com>
Change-Id: I435bf3a478dc414adf921da64c1187ea7b51d1ca